pyproject.toml
[tool.poetry]
name = "dspy-example"
version = "0.1.0"
description = "DSPy example code from article"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dspy = "^2.5.0"
openai = "^1.35.0"
anthropic = "^0.31.0"
pydantic = "^2.7.0"
python-dotenv = "^1.0.0"
requests = "^2.32.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
black = "^24.4.0"
ruff = "^0.4.0"
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"